Add: 그룹 생성페이지 색깔 클릭하여 기능구현 - 조성윤 2025-02-12 23:04 Refactor: 식당 상세페이지 리팩토링 - 조성윤 2025-02-12 23:04

+++ Prepay/app/src/main/java/com/example/prepay/data/model/dto/OrderHistory.kt
@@ -12,5 +12,5 @@ data class OrderHistory(
 
 data class OrderHistoryReq(
     val teamId: Long,
-    val storeId:Long
+    val storeId:Int
 )
